Calcula balances financieros ordenados por fecha

Calcula el balance secuencial ordenado por fecha. Ideal para estimar el balance de un conjunto de egresos/ingresos.

Uso

prices_balance(data, fecha, valor, balance_inicial = 0, alcanza = TRUE)

Argumentos

data

Datos. Deben contener como mínimo la fecha (en el cual se asigna el concepto) o número (que sirva para ordenar la secuencia del balance), así como el valor o monto asociado al egreso/ingreso. Adicionalmente puede tener una columna con los conceptos (un descriptor), el cual no se verá modificado.

fecha

Nombre de la columna con las fechas o número que sirva para ordenar los valores/montos.

valor

Nombre de la columna con los valores. Los valores/montos deben ser mayor a cero (> 0) si son ingresos, y menores a cero (< 0) si son egresos/gastos.

balance_inicial

Número indicando el balance inicial al cual se le aplican los conceptos.

alcanza

Lógico indicando si el se desea una columna indicando si el balance se mantiene positivo.

Ejemplos

prices_balance(finanzas, fecha_pago, valor)
#>     fecha_pago       concepto valor balance alcanza mes
#>  1: 2022-07-01    Venta stock  1000    1000    TRUE   7
#>  2: 2022-07-03       Arriendo  -500     500    TRUE   7
#>  3: 2022-07-05         Comida  -200     300    TRUE   7
#>  4: 2022-07-10 Gastos comunes   -70     230    TRUE   7
#>  5: 2022-07-15    Venta stock   500     730    TRUE   7
#>  6: 2022-07-28    Venta stock   200     930    TRUE   7
#>  7: 2022-07-28      Marketing  -120     810    TRUE   7
#>  8: 2022-07-28   Trabajadores  -500     310    TRUE   7
#>  9: 2022-08-01    Venta stock   100     410    TRUE   8
#> 10: 2022-08-03       Arriendo  -500     -90   FALSE   8
#> 11: 2022-08-05         Comida  -200    -290   FALSE   8
#> 12: 2022-08-10 Gastos comunes   -90    -380   FALSE   8
#> 13: 2022-08-15    Venta stock   700     320    TRUE   8
#> 14: 2022-08-28    Venta stock   300     620    TRUE   8
#> 15: 2022-08-28      Marketing  -120     500    TRUE   8
#> 16: 2022-08-28   Trabajadores  -500       0    TRUE   8